refactor: Consolidate taxon buttons in a helper
